requirements.txt
# AI洞察助手 MVP-1 依赖包
# 核心依赖
Flask==3.0.0
SQLAlchemy==2.0.23
Jinja2==3.1.2
python-dotenv==1.0.0

# 数据获取
feedparser==6.0.10
tweepy==4.14.0
requests==2.31.0
beautifulsoup4==4.12.2

# AI处理
# 使用DeepSeek API，通过requests库调用

# 数据处理
pandas==2.1.3
lxml==4.9.3

# 日志和工具
python-dateutil==2.8.2
croniter==2.0.1

# 开发工具
black==23.11.0
isort==5.12.0
